Support decimal material layout step lengths

+++ craftstudio-mes-domain/src/main/java/com/mes/domain/manufacturer/materialLayoutSpec/entity/MaterialLayoutSpecStep.java
@@ -7,9 +7,9 @@ import java.math.BigDecimal;
 @Data
 public class MaterialLayoutSpecStep {
     /**
-     * 长度阶梯上限（米），例如 1 表示 1m 内。
+     * 长度阶梯上限（米），支持小数，例如 0.1 表示 0.1m 内。
      */
-    private Integer maxLengthMeter;
+    private BigDecimal maxLengthMeter;
 
     /**
      * 内缩值（厘米）。
